article + supabase auth
imporved single article page in prep for getting the data from an api, added a link that opens the system browser over the app to th article page.

added supabase cloud integration, this service hosts the apps database that can be accessed through its api calls. use in the login/register pages, the application sends requests to the auth table in supabass to authenticate the users request/add a new user account.

that securely stores the users information with hashed passwords.