New Api Integration
replaced hard coded articles with a news article api. This uses the newsapiservice to make a call to the api and display 10 articles.

each can be clicked into a single article page, from which a user can follow a link to the full article in a browser

added description converter to truncate titles and descriptions on the articles page if they are over a certain length