Create Budget db integration
added a supabase table for budgets, which is used in the createbudgetviewmodel. A user can fill in the form and submit their information which is now sent to the cloud budgets table using the databaseservice. The users email is sent as a unique field so each user can only have one budget, and so it can be found later to read.

removed redundant files remaining from the week 8 demo starting point